Extract environment variables
